Use nightly toolchain in workflows so format action works (#22)
The configuration in rustfmt.toml appears to rely on features that are
not in the stable toolchain, as evidenced by the warnings in the 'check
formatting' stage of the Build workflow (ex:
https://github.com/mskelton/dtsfmt/actions/runs/16820506520/job/47646281792).
Switching the action to use +nightly gets the workflow back to formating
as desired.

Also run a +nightly `cargo fmt` to get everything in sync and see that
the workflow completes without warnings.